ADD Rim content type and model.
Co-authored-by: Anand Magaji <anand@eosacro.com>

Detail: Added a contentType for rim in the admin. More than the Rim, this commit
is targeted at clearing files from fields when content is deleted. Special
functions are added to the store to get all file paths and clear them from the
db. Maybe add a warning before data can be cleared.