ADD Sticker geometry and controls
Co-authored-by: Anand Magaji <anand@eosacro.com>

Detail: Added sticker geo to use World Space co-ordinates for mesh generation
and replication of mesh. Added default material to be used for the object as a
props.
Implemented Centroid to move mesh back to orignal position. Houdini facing
correction for sticker mesh.
Sub-Component in SDM to handle buffer geometry creation and manipulation.